itrust/requirements.txt
docopt
fabric3
